# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin-News-Backend-Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/news/add-news": {
            "post": {
                "tags": [
                    "add-news"
                ],
                "summary": "Add a new news",
                "description": "Create a new news. Requires JWT authentication.",
                "operationId": "a7391001ff66934e5beff30f9ccab02c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "news title",
                                        "type": "string",
                                        "example": "news 1"
                                    },
                                    "category": {
                                        "description": "category ID",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "✅ Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "news": {
                                                    "type": "string",
                                                    "example": "newsAdded"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/news/get-news": {
            "get": {
                "tags": [
                    "get-news"
                ],
                "summary": "Get all registered news who belong to the same client_id of the requester",
                "operationId": "0d7a984e52107198fb1ea422dc0650c9",
                "responses": {
                    "200": {
                        "description": "List of news",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "news": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "title": {
                                                                "type": "string",
                                                                "example": "news 1"
                                                            },
                                                            "subtitle": {
                                                                "type": "string",
                                                                "example": "subtitle 1"
                                                            },
                                                            "text": {
                                                                "type": "string",
                                                                "example": "text 1"
                                                            },
                                                            "category": {
                                                                "type": "integer",
                                                                "example": "1"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/news/update": {
            "patch": {
                "tags": [
                    "Update newss"
                ],
                "summary": "Update news data",
                "operationId": "600ea158d7d47d19e8c8e8db4b383aa2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "news_ID"
                                ],
                                "properties": {
                                    "news_ID": {
                                        "description": "news ID to update",
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "name": {
                                        "description": "Optional new news name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Optional new news email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "description": "Optional new password",
                                        "type": "string",
                                        "example": "newSecurePassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "news updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "news ID 123 updated successfully"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "news service unavailable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/news/delete": {
            "post": {
                "tags": [
                    "Delete news"
                ],
                "summary": "Delete a news by ID",
                "description": "Send the news_ID in the request body to delete the specified news.",
                "operationId": "b812c7101e7dff82d293eb90dc2c389d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "news_ID"
                                ],
                                "properties": {
                                    "news_ID": {
                                        "description": "news's ID",
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "✅ Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "news ID 123 removed"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "news service unavailable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "news not found or the news you are trying to delete doesnt belong to your same client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "news not found or the news you are trying to delete doesnt belong to your same client"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}